Move go files to the root of the repository.
