--- backend/go.mod
@@ -1,3 +1,0 @@
-module github.com/recomator/backend
-
-go 1.14
