+++ cmd/recomator-cli/main.go
@@ -16,7 +16,7 @@ limitations under the License.
 
 package main
 
-import "github.com/recomator/backend/pkg/automation"
+import "github.com/googleinterns/recomator/pkg/automation"
 
 func main() {
 	automation.ExmpleFunc()
