+++ go.mod
@@ -1,0 +1,3 @@
+module github.com/googleinterns/recomator
+
+go 1.14
