+++ pkg/automation/example.go
